fix: handler 鏈長超過 63 時溢位——Context.index 改 int32
Context.index 為 int8 且 abortIndex = math.MaxInt8/2 = 63。一旦
「全域中間件 + 路由 handler」總數超過 63，Next() 的 c.index++ 會溢位：

- 64~127 個 → index 變 -128 → index out of range [-128] panic
- 128 個以上 → 溢位後 c.index >= abortIndex 恆成立，整條鏈一個都不執行，
  卻靜默回 200 空回應（最陰險：無 log、無 panic、無錯誤碼）

實測確認（修前）：n=63 正常、n=64/70 panic、n=130 中間件與 handler
全部沒跑但回 200。

gin 用 int8 的前提是它硬性限制鏈長 ≤ 63，本專案無此檢查。改用 int32
（abortIndex 同步改 MaxInt32/2），上限遠高於任何實際鏈長；池化 Context
多 3 bytes，可忽略。

此問題在本 session 把中間件改為洋蔥模型後風險更高：預設中間件現由
Start() 自動套用，加上使用者自訂與 Group 中間件，累積鏈長更容易接近上限。

附回歸測試：63/64/130 三種鏈長都須完整執行且回應正確。

+++ pkg/context/context.go
@@ -62,7 +62,11 @@ type Context struct {
 
 	// 中間件和處理器
 	handlers []HandlerFunc
-	index    int8
+	// index 為 int32：int8 上限 127 而 abortIndex 為 63，一旦「全域中間件 +
+	// 路由 handler」總數超過 63，c.index++ 會溢位成負數，造成
+	// index out of range [-128] panic（64~127 個）或整條鏈靜默不執行、
+	// 卻回 200 空回應（128 個以上）。int32 讓上限遠高於任何實際鏈長。
+	index    int32
 	fullPath string
 
 	// 資料存儲
@@ -226,7 +230,7 @@ func (c *Context) SetHandlers(global, route []HandlerFunc) {
 // Next 執行下一個中間件
 func (c *Context) Next() {
 	c.index++
-	for c.index < int8(len(c.handlers)) {
+	for c.index < int32(len(c.handlers)) {
 		c.handlers[c.index](c)
 		c.index++
 	}
